requirements.txt
pillow
rembg
tensorflow
numpy 
matplotlib 
pathlib
protobuf